refactor: global -> support 패키지로 수정
