+++ src/main/java/kr/hhplus/be/server/domain/product/ProductSalesRank.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.domain.product;
 
 import jakarta.persistence.*;
 import kr.hhplus.be.server.domain.product.enums.RankingScope;
-import kr.hhplus.be.server.global.entity.BaseEntity;
+import kr.hhplus.be.server.support.entity.BaseEntity;
 import lombok.*;
