+++ src/main/java/kr/hhplus/be/server/domain/product/enums/RankingScope.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.domain.product.enums;
 
-import kr.hhplus.be.server.global.exception.ApiErrorCode;
-import kr.hhplus.be.server.global.exception.ApiException;
+import kr.hhplus.be.server.support.exception.ApiErrorCode;
+import kr.hhplus.be.server.support.exception.ApiException;
 
 import java.util.Arrays;
