+++ src/main/java/kr/hhplus/be/server/infra/coupon/UserCouponRepositoryImpl.java
@@ -2,8 +2,8 @@ package kr.hhplus.be.server.infra.coupon;
 
 import kr.hhplus.be.server.domain.coupon.UserCoupon;
 import kr.hhplus.be.server.domain.coupon.UserCouponRepository;
-import kr.hhplus.be.server.global.exception.ApiErrorCode;
-import kr.hhplus.be.server.global.exception.ApiException;
+import kr.hhplus.be.server.support.exception.ApiErrorCode;
+import kr.hhplus.be.server.support.exception.ApiException;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
