+++ src/main/java/kr/hhplus/be/server/support/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.global.exception;
+package kr.hhplus.be.server.support.exception;
 
-import kr.hhplus.be.server.global.ApiResponse;
+import kr.hhplus.be.server.support.ApiResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
