+++ src/test/java/kr/hhplus/be/server/domain/balance/UserBalanceServiceItTest.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.domain.balance;
 
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.domain.user.UserRepository;
-import kr.hhplus.be.server.global.exception.ApiErrorCode;
+import kr.hhplus.be.server.support.exception.ApiErrorCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
