+++ src/test/java/kr/hhplus/be/server/domain/payment/PaymentServiceTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.domain.payment;
 
-import kr.hhplus.be.server.global.exception.ApiErrorCode;
+import kr.hhplus.be.server.support.exception.ApiErrorCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
